+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/Constants.java
@@ -17,7 +17,7 @@ package io.flutter.plugins.googlemobileads;
 /** Constants used in the plugin. */
 public class Constants {
   /** Version request agent. Should be bumped alongside plugin versions. */
-  public static final String REQUEST_AGENT_PREFIX_VERSIONED = "Flutter-GMA-2.4.0";
+  public static final String REQUEST_AGENT_PREFIX_VERSIONED = "Flutter-GMA-3.0.0";
   /** Prefix for news template */
   public static final String REQUEST_AGENT_NEWS_TEMPLATE_PREFIX = "News";