+++ packages/google_mobile_ads/ios/Classes/FLTAdUtil.m
@@ -14,6 +14,7 @@
 
 #import "FLTAdUtil.h"
 #import "FLTConstants.h"
+@import webview_flutter_wkwebview;
 
 @implementation FLTAdUtil
 
@@ -45,6 +46,13 @@ static NSString *_requestAgent;
   }
   return [NSString stringWithFormat:@"%@%@%@", FLT_REQUEST_AGENT_VERSIONED,
                                     newsTemplateString, gameTemplateString];
+}
+
++ (WKWebView *)getWebView:(NSNumber *)webViewId
+    flutterPluginRegistry:(id<FlutterPluginRegistry>)flutterPluginRegistry {
+  return (WKWebView *)[FWFWebViewFlutterWKWebViewExternalAPI
+      webViewForIdentifier:webViewId.longValue
+        withPluginRegistry:flutterPluginRegistry];
 }
 
 @end