+++ packages/google_mobile_ads/lib/src/ad_instance_manager.dart
@@ -26,6 +26,10 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:google_mobile_ads/src/nativetemplates/template_type.dart';
+import 'package:google_mobile_ads/src/webview_controller_util.dart';
+import 'package:webview_flutter/webview_flutter.dart';
+import 'package:webview_flutter_android/webview_flutter_android.dart';
+import 'package:webview_flutter_wkwebview/webview_flutter_wkwebview.dart';
 
 import 'nativetemplates/native_template_font_style.dart';
 import 'nativetemplates/native_template_style.dart';
@@ -41,7 +45,8 @@ AdInstanceManager instanceManager = AdInstanceManager(
 /// Maintains access to loaded [Ad] instances and handles sending/receiving
 /// messages to platform code.
 class AdInstanceManager {
-  AdInstanceManager(String channelName)
+  AdInstanceManager(String channelName,
+      {this.webViewControllerUtil = const WebViewControllerUtil()})
       : channel = MethodChannel(
           channelName,
           StandardMethodCodec(AdMessageCodec()),
@@ -66,6 +71,7 @@ class AdInstanceManager {
 
   /// Invokes load and dispose calls.
   final MethodChannel channel;
+  final WebViewControllerUtil webViewControllerUtil;
 
   void _onAdEvent(Ad ad, String eventName, Map<dynamic, dynamic> arguments) {
     if (defaultTargetPlatform == TargetPlatform.android) {
@@ -434,14 +440,13 @@ class AdInstanceManager {
     ))!;
   }
 
-  Future<AdSize> getAdSize(Ad ad) async {
-    return (await instanceManager.channel.invokeMethod<AdSize>(
-      'getAdSize',
-      <dynamic, dynamic>{
-        'adId': adIdFor(ad),
-      },
-    ))!;
-  }
+  Future<AdSize?> getAdSize(Ad ad) =>
+      instanceManager.channel.invokeMethod<AdSize>(
+        'getAdSize',
+        <dynamic, dynamic>{
+          'adId': adIdFor(ad),
+        },
+      );
 
   /// Returns null if an invalid [adId] was passed in.
   Ad? adFor(int adId) => _loadedAds[adId];
@@ -789,6 +794,27 @@ class AdInstanceManager {
         'adUnitId': adUnitId,
       },
     );
+  }
+
+  /// Register the `WebViewController` with the GMASDK.
+  Future<void> registerWebView(WebViewController controller) {
+    return channel.invokeMethod<void>(
+      'MobileAds#registerWebView',
+      <dynamic, dynamic>{
+        'webViewId': webViewControllerUtil.webViewIdentifier(controller),
+      },
+    );
+  }
+
+  int getWebViewId(WebViewController controller) {
+    if (WebViewPlatform.instance is AndroidWebViewPlatform) {
+      return (controller.platform as AndroidWebViewController)
+          .webViewIdentifier;
+    } else if (WebViewPlatform.instance is WebKitWebViewPlatform) {
+      return (controller.platform as WebKitWebViewController).webViewIdentifier;
+    } else {
+      throw UnsupportedError('This method only supports Android and iOS.');
+    }
   }
 
   /// Send a platform message to open the ad inspector.