+++ packages/google_mobile_ads/test/mobile_ads_test.dart
@@ -21,7 +21,13 @@ import 'package:google_mobile_ads/src/ad_instance_manager.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:google_mobile_ads/src/webview_controller_util.dart';
+import 'package:mockito/annotations.dart';
+import 'package:mockito/mockito.dart';
+import 'package:webview_flutter/webview_flutter.dart';
+import 'mobile_ads_test.mocks.dart';
 
+@GenerateMocks([WebViewController, AdInstanceManager, WebViewControllerUtil])
 void main() {
   group('Mobile Ads', () {
     TestWidgetsFlutterBinding.ensureInitialized();
@@ -34,6 +40,8 @@ void main() {
     });
 
     setUp(() async {
+      instanceManager =
+          AdInstanceManager('plugins.flutter.io/google_mobile_ads');
       log.clear();
       MethodChannel(
         'plugins.flutter.io/google_mobile_ads',
@@ -59,6 +67,8 @@ void main() {
           case 'MobileAds#getVersionString':
             return Future<String>.value('Test-SDK-Version');
           case 'MobileAds#updateRequestConfiguration':
+            return null;
+          case 'MobileAds#registerWebView':
             return null;
           case 'MobileAds#getRequestConfiguration':
             return RequestConfiguration(
@@ -152,6 +162,46 @@ void main() {
       expect(status.latency, 0);
     });
 
+    test('$MobileAds.registerWebView', () async {
+      instanceManager = MockAdInstanceManager();
+      final webView = MockWebViewController();
+      when(instanceManager.registerWebView(webView))
+          .thenAnswer((realInvocation) => Future.value());
+      await MobileAds.instance.registerWebView(webView);
+
+      verify(instanceManager.registerWebView(webView));
+    });
+
+    test('$MobileAds.registerWebView android', () async {
+      debugDefaultTargetPlatformOverride = TargetPlatform.android;
+      final mockWebViewControllerUtil = MockWebViewControllerUtil();
+      when(mockWebViewControllerUtil.webViewIdentifier(any)).thenReturn(1);
+      instanceManager = AdInstanceManager(
+          'plugins.flutter.io/google_mobile_ads',
+          webViewControllerUtil: mockWebViewControllerUtil);
+      final webView = MockWebViewController();
+      await MobileAds.instance.registerWebView(webView);
+
+      expect(log, <Matcher>[
+        isMethodCall('MobileAds#registerWebView', arguments: {'webViewId': 1})
+      ]);
+    });
+
+    test('$MobileAds.registerWebView iOS', () async {
+      debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
+      final mockWebViewControllerUtil = MockWebViewControllerUtil();
+      when(mockWebViewControllerUtil.webViewIdentifier(any)).thenReturn(1);
+      instanceManager = AdInstanceManager(
+          'plugins.flutter.io/google_mobile_ads',
+          webViewControllerUtil: mockWebViewControllerUtil);
+      final webView = MockWebViewController();
+      await MobileAds.instance.registerWebView(webView);
+
+      expect(log, <Matcher>[
+        isMethodCall('MobileAds#registerWebView', arguments: {'webViewId': 1})
+      ]);
+    });
+
     test('$MobileAds.openAdInspector success', () async {
       MethodChannel(
         'plugins.flutter.io/google_mobile_ads',
@@ -190,6 +240,8 @@ void main() {
     });
 
     test('$MobileAds.setSameAppKeyEnabled', () async {
+      debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
+
       await MobileAds.instance.setSameAppKeyEnabled(true);
 
       expect(log, <Matcher>[