+++ packages/google_mobile_ads/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 3.0.0
+* Adds support for `MobileAds.registerWebView()`. This API supports in-app ad monetization for
+  `WebView`s. You can read more in the [android](https://developers.google.com/admob/android/webview)
+  or [iOS](https://developers.google.com/admob/ios/webview) documentation.
+  This plugin now depends on [webview_flutter](https://pub.dev/packages/webview_flutter), 
+  [webview_flutter_wkwebview](https://pub.dev/packages/webview_flutter_wkwebview), and 
+  [webview_flutter_android](https://pub.dev/packages/webview_flutter_android)
+* Updates Android GMA dependency to 22.0.0 and iOS dependency to 10.4.0
+* Fixes https://github.com/googleads/googleads-mobile-flutter/issues/700
+
 ## 2.4.0
 * Adds support for native templates, which are predefined layouts for native ads.
   * `NativeAd` has a new optional parameter, `nativeTemplateStyle` of type `NativeTemplateStyle`.